src/main/java/statemachine/states/states/LoadTeamState.java
package statemachine.states.states;


import leagueobjectmodel.ILeagueModel;
import leagueobjectmodel.LeagueObjectModelAbstractFactory;
import org.apache.log4j.Logger;
import statemachine.loadteam.ILoadTeam;
import statemachine.loadteam.LoadTeamAbstractFactory;
import statemachine.states.StateMachine;

public class LoadTeamState implements ITransition {
    StateMachine stateMachine;
    ILeagueModel currentLeague;
    ILoadTeam parser;
    final static Logger logger = Logger.getLogger(LoadTeamState.class);
    public LoadTeamState(StateMachine stateMachine) {
        logger.info("Initializing LoadTeam State");
        this.stateMachine = stateMachine;
        currentLeague = LeagueObjectModelAbstractFactory.getInstance().getLeague();
    }

    public StateMachine getStateMachine() {
        return stateMachine;
    }

    public void setStateMachine(StateMachine stateMachine) {
        this.stateMachine = stateMachine;
    }

    @Override
    public void entry() throws Exception {
        parser = LoadTeamAbstractFactory.instance().createLoadTeam();
        currentLeague = parser.getData();
        task();
    }

    @Override

    public void task() throws Exception {
        if (currentLeague == null){
            task();
        }
        else {
            exit();
        }
    }

    @Override
    public void exit() throws Exception {
        stateMachine.getUpdateStateValue().updatePlayerSeasonChoiceStateValue(stateMachine,currentLeague);
        stateMachine.setCurrentState(stateMachine.teamLoaded());
        stateMachine.getCurrentState().entry();
    }
}
